MNT: Reorganize.
